% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_split.R
\name{codon_split}
\alias{codon_split}
\title{Split RNA strand into codons}
\usage{
codon_split(rna, start = 1)
}
\arguments{
\item{rna}{An RNA stand}

\item{start}{The starting position}
}
\value{
A list of codons
}
\description{
Split RNA strand into codons
}
