% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abundance.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Plot abundance of each amino acid in a peptide}
\usage{
plot_abundance(peptide)
}
\arguments{
\item{peptide}{A peptide}
}
\value{
An bar chart of amino acid content
}
\description{
Plot abundance of each amino acid in a peptide
}
