% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Create a random DNA strand}
\usage{
random_dna(dna_length)
}
\arguments{
\item{dna_length}{The length of a random DNA strand}
}
\value{
A randomly generated DNA strand
}
\description{
Create a random DNA strand
}
