% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{Transcribe a DNA strand into RNA}
\usage{
transcribe(dna)
}
\arguments{
\item{dna}{A DNA strand}
}
\value{
A RNA strand
}
\description{
Transcribe a DNA strand into RNA
}
