% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate a list of codons into a peptide}
\usage{
translate(codons)
}
\arguments{
\item{codons}{A list of codons}
}
\value{
A peptide
}
\description{
Translate a list of codons into a peptide
}
