% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{Transcript a list of codons into a peptide}
\usage{
transcribe(codons)
}
\arguments{
\item{codons}{A list of codons}
}
\value{
A peptide
}
\description{
Transcript a list of codons into a peptide
}
