% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translates a DNA strand into RNA}
\usage{
translate(dna)
}
\arguments{
\item{dna}{A DNA strand}
}
\value{
A RNA strand
}
\description{
Translates a DNA strand into RNA
}
